Notebook + 3 modules
As a suggestion I started building (my) final notebook. Feel free to add yours to it. 
The modules to generate the train, test and optionally the eval dataset, the module for the train-test-splitting and the module for the LightGBM are uploaded. The final_notebook contains explanations how to use them.